QR-Generator-mail-sender/requirements.txt
qrcode[pil]
pandas
openpyxl
pillow
python-dotenv